+++ routes/web.php
@@ -9,6 +9,8 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [LandingController::class, 'index'])->name('landing');
 
 Route::prefix('booking')->name('booking.')->group(function () {
+    Route::get('/data-pemesan', [BookingController::class, 'customer'])->name('customer');
+    Route::post('/data-pemesan', [BookingController::class, 'storeCustomer'])->name('customer.store');
     Route::get('/', [BookingController::class, 'create'])->name('create');
     Route::get('/availability', [BookingController::class, 'availability'])->name('availability');
     Route::get('/payment', [BookingController::class, 'payment'])->name('payment');